requirement.txt
asgiref==3.2.7
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
dotmap==1.3.13
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
